// src/app/shared/interfaces/products/IProduct.ts
import { IProductCategory } from './IProductCategory';
import { ISubcategory } from './ISubcategory';

export interface IProduct {
  sold: number;
  images: string[];
  subcategory: ISubcategory[];
  ratingsQuantity: number;
  _id: string;
  title: string;
  slug: string;
  description: string;
  quantity: number;
  price: number;
  imageCover: string;
  category: IProductCategory;
  brand: IProductCategory;
  ratingsAverage: number;
  createdAt: string;
  updatedAt: string;
  id: string;
  priceAfterDiscount?: number;
  availableColors?: string[];
}